+++ src/main/java/org/github/saphyra/skyxplore/ship/EquippedShipConstants.java
@@ -1,0 +1,16 @@
+package org.github.saphyra.skyxplore.ship;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+final class EquippedShipConstants {
+    static final String DEFENSE_SLOT_NAME = "defense";
+    static final String WEAPON_SLOT_NAME = "weapon";
+    static final String CONNECTOR_SLOT_NAME = "connector";
+
+    static final String FRONT_SLOT_NAME = "front";
+    static final String BACK_SLOT_NAME = "back";
+    static final String LEFT_SLOT_NAME = "left";
+    static final String RIGHT_SLOT_NAME = "right";
+}
