+++ src/test/java/org/github/saphyra/skyxplore/ship/ShipControllerTest.java
@@ -1,18 +1,17 @@
 package org.github.saphyra.skyxplore.ship;
 
-import static org.junit.Assert.assertEquals;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-
+import org.github.saphyra.skyxplore.ship.domain.EquipRequest;
+import org.github.saphyra.skyxplore.ship.domain.ShipView;
+import org.github.saphyra.skyxplore.ship.domain.UnequipRequest;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 
-import org.github.saphyra.skyxplore.ship.domain.EquipRequest;
-import org.github.saphyra.skyxplore.ship.domain.UnequipRequest;
-import org.github.saphyra.skyxplore.ship.domain.ShipView;
+import static org.junit.Assert.assertEquals;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
 public class ShipControllerTest {
@@ -22,7 +21,16 @@ public class ShipControllerTest {
     private static final String SLOT = "slot";
 
     @Mock
-    private EquippedShipFacade equippedShipFacade;
+    private EquipService equipService;
+
+    @Mock
+    private EquipShipService equipShipService;
+
+    @Mock
+    private ShipQueryService shipQueryService;
+
+    @Mock
+    private UnequipService unequipService;
 
     @InjectMocks
     private ShipController underTest;
@@ -34,7 +42,7 @@ public class ShipControllerTest {
         //WHEN
         underTest.equip(equipRequest, CHARACTER_ID);
         //THEN
-        verify(equippedShipFacade).equip(equipRequest, CHARACTER_ID);
+        verify(equipService).equip(equipRequest, CHARACTER_ID);
     }
 
     @Test
@@ -42,14 +50,14 @@ public class ShipControllerTest {
         //WHEN
         underTest.equipShip(CHARACTER_ID, ITEM_ID);
         //THEN
-        verify(equippedShipFacade).equipShip(CHARACTER_ID, ITEM_ID);
+        verify(equipShipService).equipShip(CHARACTER_ID, ITEM_ID);
     }
 
     @Test
     public void testGetShipDataShouldCallFacadeAndReturnView() {
         //GIVEN
         ShipView shipView = ShipView.builder().build();
-        when(equippedShipFacade.getShipData(CHARACTER_ID)).thenReturn(shipView);
+        when(shipQueryService.getShipData(CHARACTER_ID)).thenReturn(shipView);
         //WHEN
         ShipView result = underTest.getShipData(CHARACTER_ID);
         //THEN
@@ -63,6 +71,6 @@ public class ShipControllerTest {
         //WHEN
         underTest.unequip(unequipRequest, CHARACTER_ID);
         //THEN
-        verify(equippedShipFacade).unequip(unequipRequest, CHARACTER_ID);
+        verify(unequipService).unequip(unequipRequest, CHARACTER_ID);
     }
 }
